requirements.txt
# AWS SageMaker and ML dependencies
sagemaker>=2.190.0
boto3>=1.26.0
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0
matplotlib>=3.5.0
seaborn>=0.11.0
opencv-python>=4.5.0
Pillow>=8.3.0
torch>=1.11.0
torchvision>=0.12.0
tensorflow>=2.8.0
jupyter>=1.0.0
jupyterlab>=3.4.0
ipywidgets>=7.6.0

# Data processing
opencv-python-headless>=4.5.0
albumentations>=1.1.0
imgaug>=0.4.0

# Visualization and monitoring
plotly>=5.0.0
wandb>=0.12.0
tensorboard>=2.8.0

# Development tools
black>=22.0.0
flake8>=4.0.0
pytest>=7.0.0